Add files via upload
The Scientific Calculator In Python project developed using Python. This is a GUI-base project which is very easy to understand and use. Also, this project makes a convenient way for the user to gain an idea of how to perform numerical calculations.

The project contains the numbers, operators, and signs like on a normal calculator. So, the user can enter any number or can just click on the numbers they want in the calculations. Hence, the user can use a simple calculator from this application.

This scientific calculator project is in Python. Talking about the features of this system, this python application is designed to calculate entered numbers to any operations and it is also capable of handling all types of exceptions. Also, the design of this system is pretty simple so that the user won’t get any difficulties while working on it.